[SERVICE-526] Move tray icon file
