+++ src/provider/model/FinTrayIcon.ts
