+++ src/provider/model/TrayIcon.ts
